pyproject.toml
[project]
name = "observe_me"
version = "1.0.0-b.4"
description = "Fastapi to control the HDAS remotly"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.10.6",
    "pydantic-settings>=2.12.0",
    "python-keycloak>=5.3.0",
    "uvicorn>=0.34.0",
    "fastapi>=0.115.12",
    "rich >=14.1.0",
    "psutil>=7.0.0",
    "py-cpuinfo>=9.0.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pyjwt>=2.12.1",
    "aws-opentelemetry-distro>=0.10.1",
    "opentelemetry-exporter-otlp>=1.25,<2",
    "opentelemetry-sdk",
]

[dependency-groups]
dev = [
    "pre-commit==4.5.1",
    "pyrefly==0.47.0",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cookies==0.7.0",
    "pytest-cov==7.0.0",
    "pytest-xdist==3.8.0",
    "ruff==0.14.11",
    "ty==0.0.10",
    "python-semantic-release>=10.5.2",
    "nodejs>=0.1.1", # required for pyright
    "pyright>=1.1.400",
    "build",
    "twine",
]
docs = [
    "mkdocs==1.6.1",
    "mkdocs-material==9.7.0",
    "mkdocstrings-python==1.19.0",
    "mkdocs-awesome-pages-plugin==2.9.3",
]


[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[tool.uv]
package = true

[tool.semantic_release]
build_command = "uv build"
changelog_file = "CHANGELOG.md"
version_variables = [
    "pyproject.toml:version",
    "src/observe_me/config/__init__.py:__version__",
]

[tool.semantic_release.remote]
type = "github"
token = {env="GH_TOKEN"}

[tool.semantic_release.branches.master]
match = "master"
prerelease = false

[tool.semantic_release.branches.develop]
match = "develop"
prerelease = true
prerelease_token = "b"

[tool.ruff]
preview = true
target-version = "py313"
line-length = 120
indent-width = 4
extend-exclude = [".venv", "build", "dist", "notebooks_extra", "tests", "src/__main__.py"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = "dynamic"


[tool.ruff.lint]
select = [
    "A",   # flake8-builtins
    "ANN", # flake8-annotations
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "C90", # mccabe
    "D",   # pydocstyle
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "PGH", # pygrep-hooks
    "RUF", # ruff-specific rules
    "S",   # flake8-bandit
    "SIM", # flake8-simplify
    "T10", # flake8-debugger
    "T20", # flake8-print
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
    "YTT", # flake8-2020
]
ignore = ["ANN401"]


[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.ruff.lint.isort]
known-first-party = ["orion"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]


[tool.ruff.lint.per-file-ignores]
"**/notebooks_extra/**" = ["ALL"]
"**.ipynb" = ["E501", "D100", "D103", "D101", "E402", "I002", "T201"]
"**/__init__.py" = ["D104", "F401"]
"tests/**" = ["S101", "D103", "D100", "D104", "SLF001", "B018"]


[tool.ty.rules]


[tool.ty.environment]
python = "./.venv"
python-version = "3.13"
python-platform = "linux"
root = ["."]
extra-paths = []


[tool.ty.src]
respect-ignore-files = false
include = ["src"]
exclude = ["**/node_modules", "**/__pycache__", "**/*venv/**/*", "./build", "**/__main__.py"]


[tool.pyrefly]
use-ignore-files = false
project-includes = ["src"]
project-excludes = [
    "**/node_modules",
    "**/__pycache__",
    "**/*venv/**/*",
    "**/tests", "**/__main__.py"
]

search-path = [
    "src",
    ".venv/lib/python3.13/site-packages",
]
python-platform = "linux"
python-version = "3.13"
python-interpreter-path = ".venv/bin/python3"


[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-n",
    "auto",
    "--cov=src",
    "--cov-report=xml",
    "--cov-report=json",
    "--cov-report=html",
    "--cov-report=term-missing",
]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
python_files = "test_*.py tests_*.py *_test.py"
python_classes = "*Test Test* Describe*"
python_functions = "test_* it_* should_* scenario_*"

[tool.coverage.run]
omit = [
    "src/__main__.py",
]

[tool.coverage.report]
fail_under = 80          # número entero o float (ej: 85.5)
#show_missing = true
#skip_covered = true
precision = 2            # cuántos decimales mostrar
#sort = "Cover"           # o "Miss", "Name", etc.
#exclude_also = [
#    "if TYPE_CHECKING:",
#    "raise NotImplementedError",
#]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.hatch.build.targets.wheel]
packages = ["src/das_control"]

[tool.hatch.build]
exclude = [
 ".venv",
 ".cache",
 "__pycache__",
 "*.pyc",
 ".git",
 ".github",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
